webapp/api/requirements.txt
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
psycopg2-binary==2.9.9
boto3==1.34.34